+++ scripts/git-subtree/webview/add.ps1
@@ -1,0 +1,5 @@
+git subtree add `
+    -P workspace/plugins/webview `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-webview.git `
+    dev `
+    --squash