+++ scripts/git-subtree/webview/push.ps1
@@ -1,0 +1,4 @@
+git subtree push `
+    -P workspace/plugins/webview `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-webview.git `
+    dev